Move "LocalFeedItem" to separate file
