+++ FeedLoader/FeedCache/FeedStore.swift
@@ -14,16 +14,3 @@ protocol FeedStore {
     func insertCache(with items: [LocalFeedItem], timestamp: Date, completion: @escaping Completion)
 }
 
-struct LocalFeedItem: Equatable {
-    let id: UUID
-    let description: String?
-    let location: String?
-    let imageUrl: URL
-    
-    init(id: UUID, description: String?, location: String?, imageUrl: URL) {
-        self.id = id
-        self.description = description
-        self.location = location
-        self.imageUrl = imageUrl
-    }
-}
